requirements.txt
flask==2.3.3
openai==1.3.0
psycopg2-binary==2.9.7
google-cloud-secret-manager==2.16.4
